Add option to writeup generator to only check, not write

+++ src/python/generate_writeups.py
@@ -1,6 +1,7 @@
 import os
 import json
 import shutil
+from argparse import ArgumentParser
 
 from typing import Dict, Optional, Tuple
 
@@ -23,19 +24,19 @@ def read_if_file(cwd: str, entry: str) -> str:
     return entry
 
 
-def render_spec(template: str, cwd: str, spec: Dict[str, str]) -> Tuple[str, str]:
+def render_spec(template: str, cwd: str, ctf_path: str, spec: Dict[str, str]) -> Tuple[str, str]:
     format_args = {
-        "title": spec["title"],
+        "title": os.path.join(ctf_path, spec["name"]),
         "author": format_if_present("> {}\n", spec.get("author")),
         "points": spec["points"],
-        "question": read_if_file(cwd, spec["question"]),
-        "solution": read_if_file(cwd, spec["solution"]),
+        "question": read_if_file(cwd, spec["question"]).strip(),
+        "solution": read_if_file(cwd, spec["solution"]).strip(),
         "flag": spec["flag"],
     }
-    return OUTPUT_FILE_NAME_FORMAT.format(**format_args), template.format(**format_args)
+    return OUTPUT_FILE_NAME_FORMAT.format(title=spec["name"]), template.format(**format_args)
 
 
-def main() -> None:
+def main(check: bool) -> None:
     current_dir = os.path.dirname(__file__)
     top_level_dir = os.path.join(current_dir, "..", "..")
     data_dir = os.path.join(top_level_dir, "src", "data")
@@ -44,12 +45,15 @@ def main() -> None:
     with open(os.path.join(data_dir, TEMPLATE_FILE_NAME)) as f:
         template = f.read()
 
-    shutil.rmtree(output_dir, ignore_errors=True)
+    if not check:
+        shutil.rmtree(output_dir, ignore_errors=True)
 
+    retval = 0
     for root, source_dirs, _ in os.walk(data_dir):
         for source_dir in source_dirs:
             source_dir = os.path.join(root, source_dir)
             source_spec = os.path.join(source_dir, SPEC_FILE_NAME)
+            ctf_dir = os.path.relpath(root, data_dir)
 
             if not os.path.exists(source_spec):
                 # Not a bottom-level writeup directory.
@@ -57,18 +61,39 @@ def main() -> None:
 
             with open(source_spec, "r") as f:
                 output_file, output_file_contents = render_spec(
-                    template, source_dir, json.loads(f.read())
+                    template, source_dir, ctf_dir, json.loads(f.read())
                 )
 
-            output_file_dir = os.path.join(output_dir, os.path.relpath(root, data_dir))
+            output_file_dir = os.path.join(output_dir, ctf_dir)
             os.makedirs(output_file_dir, exist_ok=True)
 
             output_file = os.path.join(output_file_dir, output_file)
-            with open(output_file, "w") as f:
-                f.write(output_file_contents)
+            if check:
+                failed = False
+                if os.path.exists(output_file):
+                    with open(output_file, "r") as f:
+                        failed = (f.read() != output_file_contents)
+                else:
+                    failed = True
+                if failed:
+                    print(f"File `{os.path.relpath(output_file, top_level_dir)} needs to be regenerated`")
+                    retval = 1
+            else:
+                with open(output_file, "w") as f:
+                    f.write(output_file_contents)
 
-            print(f"Rendered `{os.path.relpath(output_file, top_level_dir)}`")
+            if not check:
+                print(f"Rendered `{os.path.relpath(output_file, top_level_dir)}`")
+
+    if retval:
+        print()
+        print("Regenerate files with `python3 src/python/generate_writeups.py`")
+    return retval
 
 
 if __name__ == "__main__":
-    main()
+    parser = ArgumentParser(description="Generate writeup documents")
+    parser.add_argument("filenames", nargs="*", action="append")
+    parser.add_argument("--check", dest="check", action="store_true")
+    args = parser.parse_args()
+    raise SystemExit(main(args.check))
